Splitted QuadTree to class patterns.
